moved dead commands to old commands folder
